Fixed mitochondria tests and error handling in mitochondria
Only took about 4 hours.

There was an issue with async in the tests, causing it to wait for other
processes to release the lock on objects before trying to change them,
which led to all kinds of problems.
