+++ src/mitochondria/__test__/run.test.ts
@@ -1,0 +1,17 @@
+describe('mitochondria', () => {
+  describe('transactions', () => {
+    require('./transaction');
+  });
+  describe('balances', () => {
+    require('./balances');
+  });
+  describe('company', () => {
+    require('./company');
+  });
+  describe('auth', () => {
+    require('./auth');
+  });
+});
+
+// All files must be modules blablbla
+export default null;
